requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
anthropic==0.43.0
supabase==2.10.0
python-dotenv==1.0.1
simpleeval==0.9.13
fpdf2==2.8.1
Pillow==11.1.0
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.0
